# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMM API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "CMM env"
        }
    ],
    "paths": {
...
        "/responsables": {
            "get": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "getResponsable",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "paginate",
                        "required": false,
                        "schema": {
                            "title": "Paginate",
                            "description": "The Paginate data",
                            "type": "boolean",
                            "example": "true"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string",
                            "example": "id"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string",
                            "example": "desc"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Sort order field",
                        "schema": {
                            "title": "perPage",
                            "description": "The unique identifier of a curso resource",
                            "type": "number",
                            "default": "10"
                        }
                    },
                    {
                        "name": "dataSearch",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "Search data",
                            "type": "string"
                        }
                    },
                    {
                        "name": "dataFilter",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Responsables all."
                    },
                    "default": {
                        "description": "error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Creates a new Responsable",
                "description": "Creates a new Responsable",
                "operationId": "6b5665cb636c2161fd32ad011e93a77d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Responsable"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The Responsable resource created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/responsables/{id}": {
            "get": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Get Responsable information",
                "description": "Returns Responsable data",
                "operationId": "getResponsableById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Responsable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Responsable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Update existing Responsable",
                "description": "Returns updated Responsable data",
                "operationId": "updateResponsable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Responsable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Responsable"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Responsable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Responsable"
                ],
                "summary": "Delete existing Responsable",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteResponsable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Responsable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Responsable": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "properties": {
                            "name": {
                                "description": "The Responsable name",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "last_name": {
                                "description": "The Responsable last_name",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "document_number": {
                                "description": "The Responsable document_number",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "email": {
                                "description": "The Responsable email",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "phone_number": {
                                "description": "The Responsable phone_number",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "user_created_id": {
                                "description": "The Responsable crete",
                                "required": [
                                    "true"
                                ],
                                "type": "number",
                                "example": 1
                            },
                            "beneficiary_id": {
                                "description": "The Responsable beneficiary",
                                "required": [
                                    "true"
                                ],
                                "type": "number",
                                "example": 1
                            },
                            "user_updated_id": {
                                "description": "The Responsable update",
                                "required": [
                                    "true"
                                ],
                                "type": "number",
                                "example": 1
                            }
                        }
                    }
                ]
            },
            "User": {
                "properties": {
                    "name": {
                        "description": "The user name",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "last_name": {
                        "description": "The Users password",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "email": {
                        "description": "The Users email",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "roles": {
                        "description": "The Product attribute_types",
                        "required": [
                            "false"
                        ],
                        "type": "array",
                        "items": {
                            "properties": {
                                "role_id": {
                                    "type": "number"
                                },
                                "branch_office_id": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "password": {
                        "description": "The Users password",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "user_created_id": {
                        "description": "The Users crete",
                        "required": [
                            "true"
                        ],
                        "type": "number",
                        "example": 1
                    },
                    "user_updated_id": {
                        "description": "The Users update",
                        "required": [
                            "true"
                        ],
                        "type": "number",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}